+++ src/init.cpp
@@ -17,6 +17,7 @@ using namespace boost;
 void rescanfornames();
 
 CWallet* pwalletMain;
+string walletPath;
 
 //////////////////////////////////////////////////////////////////////////////
 //
@@ -25,7 +26,7 @@ CWallet* pwalletMain;
 
 void ExitTimeout(void* parg)
 {
-#ifdef __WXMSW__
+#ifdef _WIN32
     MilliSleep(5000);
     ExitProcess(0);
 #endif
@@ -136,10 +137,9 @@ bool AppInit2(int argc, char* argv[])
     // Disable confusing "helpful" text message on abort, ctrl-c
     _set_abort_behavior(0, _WRITE_ABORT_MSG | _CALL_REPORTFAULT);
 #endif
-#ifndef __WXMSW__
+#ifndef _WIN32
     umask(077);
-#endif
-#ifndef __WXMSW__
+
     // Clean shutdown on SIGTERM
     struct sigaction sa;
     sa.sa_handler = HandleSIGTERM;
@@ -239,7 +239,7 @@ bool AppInit2(int argc, char* argv[])
         exit(ret);
     }
 
-#ifndef __WXMSW__
+#ifndef _WIN32
     if (fDaemon)
     {
         // Daemonize
@@ -406,9 +406,14 @@ bool AppInit2(int argc, char* argv[])
     printf("Loading wallet...\n");
     nStart = GetTimeMillis();
     bool fFirstRun;
-    pwalletMain = new CWallet("wallet.dat");
+    string argWalletPath = GetArg("-walletpath", "wallet.dat");
+    boost::filesystem::path pathWalletFile(argWalletPath);
+    walletPath = pathWalletFile.string();
+    
+    pwalletMain = new CWallet(walletPath);
     if (!pwalletMain->LoadWallet(fFirstRun))
-        strErrors += _("Error loading wallet.dat      \n");
+      strErrors += "Error loading " + argWalletPath + "      \n";
+    
     printf(" wallet      %15"PRI64d"ms\n", GetTimeMillis() - nStart);
 
     RegisterWallet(pwalletMain);
@@ -433,7 +438,7 @@ bool AppInit2(int argc, char* argv[])
         pindexRescan = pindexGenesisBlock;
     else
     {
-        CWalletDB walletdb("wallet.dat");
+        CWalletDB walletdb(walletPath);
         CBlockLocator locator;
         if (walletdb.ReadBestBlock(locator))
             pindexRescan = locator.GetBlockIndex();
@@ -594,6 +599,7 @@ std::string HelpMessage()
     std::string strUsage = std::string(_("Options:\n")) +
         "  -conf=<file>     \t\t  " + _("Specify configuration file (default: namecoin.conf)\n") +
         "  -pid=<file>      \t\t  " + _("Specify pid file (default: namecoind.pid)\n") +
+        "  -walletpath=<file> \t  " + _("Specify the wallet filename (default: wallet.dat)") + "\n" +
         "  -gen             \t\t  " + _("Generate coins\n") +
         "  -gen=0           \t\t  " + _("Don't generate coins\n") +
         "  -min             \t\t  " + _("Start minimized\n") +